added tests folder

+++ Project/models/order.py
@@ -1,5 +1,5 @@
 from . import db
-from sqlalchemy import Integer, Float, DateTime, ForeignKey, DECIMAL
+from sqlalchemy import Integer, Float, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 
 class Order(db.Model):
@@ -9,7 +9,7 @@ class Order(db.Model):
     user_id = db.mapped_column(Integer, ForeignKey('user.id'))
     date_created = db.mapped_column(DateTime)
     date_completed = db.mapped_column(DateTime, nullable=True)
-    amount = db.mapped_column(DECIMAL(6, 2), nullable=False)
+    amount = db.mapped_column(Float, nullable=False)
     
     user = relationship("User", back_populates="orders")
     items = relationship("Orderbook", back_populates="order")
